requirements.txt
torch
torchvision
opencv-python
numpy
pillow
matplotlib
scikit-image
pyyaml
tqdm
